backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
celery==5.3.4
redis==5.0.1

# 音声処理
librosa==0.10.1
soundfile==0.12.1
scipy==1.11.4
numpy==1.26.2
praat-parselmouth==0.4.3
pydub==0.25.1

# 機械学習
torch==2.1.1
scikit-learn==1.3.2

# エクスポート
openpyxl==3.1.2
reportlab==4.0.7
matplotlib==3.8.2
Pillow==10.1.0

# テスト
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# その他
python-dotenv==1.0.0
